🍻 refactor: useEffect, useState  커스텀훅으로 정리
